CI-CD/pf-cicd-agent/pyproject.toml
# =============================================================================
# PF-CORE CI/CD Automation Agent
# =============================================================================
# Python project configuration for the agentic CI/CD orchestration system
# =============================================================================

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pf-cicd-agent"
version = "1.0.0"
description = "PF-CORE CI/CD Automation Agent using Claude Agent SDK"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Platform Foundation Core Holdings" }
]
keywords = ["ci-cd", "automation", "agent", "claude", "devops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "anthropic>=0.40.0",
    "PyGithub>=2.1.0",
    "python-digitalocean>=1.17.0",
    "supabase>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "jsonschema>=4.20.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.2",
    "paramiko>=3.4.0",
    "httpx>=0.25.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "types-PyYAML>=6.0.0",
    "types-paramiko>=3.4.0",
]

[project.scripts]
pf-cicd = "pf_cicd_agent.cli:main"

[project.urls]
"Homepage" = "https://github.com/ajrmooreuk/PF-Prototype-Shared"
"Documentation" = "https://github.com/ajrmooreuk/PF-Prototype-Shared/tree/main/CI-CD/pf-cicd-agent/docs"

[tool.hatch.build.targets.wheel]
packages = ["src/pf_cicd_agent"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/templates",
    "/schemas",
]

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["pf_cicd_agent"]

[tool.ruff.per-file-ignores]
"tests/**/*" = ["ARG001", "S101"]

# =============================================================================
# Black Configuration
# =============================================================================
[tool.black]
target-version = ["py311"]
line-length = 100
include = '\.pyi?$'

# =============================================================================
# MyPy Configuration
# =============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "digitalocean.*",
    "github.*",
    "supabase.*",
    "paramiko.*",
]
ignore_missing_imports = true

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/pf_cicd_agent"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true
